Fixed ui

+++ Assets/Scripts/Stats/EnemySO.cs
@@ -5,6 +5,6 @@ public class EnemySO : ScriptableObject
 {
     public float movementSpeed = 5f;
     public int maxHealth = 10;
-    public float attackDamage = 10f;
-    public float moneyAfterDeath = 10.0f;
+    public int attackDamage = 10;
+    public int moneyAfterDeath = 10;
 }
